Fix 0 timer

+++ main.py
@@ -7,11 +7,12 @@ def submit():
         minuitsms=int(minuits.get())
         secondsms=int(minuits.get())
         final=hrsms*3600+minuitsms*60+secondsms
-        system(f"shutdown -s -t {final}")
-        hrs.set(int(0))
-        minuits.set(int(0))
-        seconds.set(int(0)) 
-        cancel_btn.grid(row=2,sticky='e')
+        if((hrsms or minuitsms or secondsms)):
+            system(f"shutdown -s -t {final}")
+            hrs.set(int(0))
+            minuits.set(int(0))
+            seconds.set(int(0)) 
+            cancel_btn.grid(row=2,sticky='e')
     except ValueError:
         hrs.set(int(0))
         minuits.set(int(0))
